Refactored discrete_distribution's iterator support a bit
iterator_types and const_iterator_types are now no longer templates, which
simplifies things, and the iterators are nested classes, which allows easy
instantiation of discrete_distribution including iterators. The actual
implementation is inherited from iterator_base, which is (as before) a part
of the discrete_distribution_details namespace. This design now mimics the
design of the iteration support of vector_distribution.