Fixed zero_level to not cause signed integer overflows, which are undefined behaviour in C++
Previously, zero_level was set to the smallest possible value, meaning that -zero_level caused
an overflow because the range of signed integer types is larger (by one) in the negative integers
than in the positive integers. Now, zero_level is the smallest value plus one, this should avoid
the issue.