Fixed technically invalid memory access in levels_nonzero_remove().
The previous code accessed the previously last element of a vector after pop_back().
While UBSAN correctly flagged the previous behaviour as an out-of-bounds access, it's
very unlikely that this could have caused problems in pratice. The value in question
was a pointer type, and the vector had hence (for the lack of a destructor) no reason
to scribble over the element during pop_back(). There also weren't any other vector
operations performed between pop_back() and the invalid access that could have caused
the element to get corrupted.